Move native program entrypoint out of instruction_processor_utils
