+++ sdk/src/lib.rs
@@ -6,6 +6,7 @@ pub mod account_utils;
 pub mod bpf_loader;
 pub mod clock;
 pub mod commitment_config;
+pub mod entrypoint_native;
 pub mod epoch_schedule;
 pub mod fee_calculator;
 pub mod hash;
